saved cars page - added a gap between the "add to basket" and "removed from saved" buttons.
car details page - made the input box in the comment section be within the container.

account settings page - made it so the red color is consistent and added hover css to the "save" and "update password" buttons.

compare cars page - got rid of the black/white lines on the left-hand-side except the line at the very top right underneath the "select product".

order details admin page - changed the navbar so it is consistent with the rest of the website, changed the css of the "remove order" button.

admin list page - made the "order status" dropdown more appealing, put the search box and search button in the middle of its div section, made the "view" button red and added hover css to it.

productslistadmin page - make the "edit" and "delete" button consistent with each other so same width height and colour, and make the gap between each container bigger.

customer list page - added a gap between the containers, and made the red colour consistent with the rest of the website.

customer amendments page - made the red colour consistent with the rest of the website including the "update" button, and when hovering over each user information, and the line at the top just above "USER INFORMATION"